% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfStats.R
\name{df_all_stats}
\alias{df_all_stats}
\title{Calculates all statistics over all numeric columns in the given data frame.}
\usage{
df_all_stats(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
An object containing all stats for the given data frame.
}
\description{
Calculates all statistics over all numeric columns in the given data frame.
}
\examples{
df_stats = df_all_stats(test_data)
print(df_stats)
}
