% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfStats.R
\name{df_average}
\alias{df_average}
\title{Calculates the average over all numeric columns in the given data frame.}
\usage{
df_average(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
The average of the data frame.
}
\description{
Calculates the average over all numeric columns in the given data frame.
}
\examples{
average = df_average(test_data)
print(average)
}
