% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfStats.R
\name{df_mean}
\alias{df_mean}
\title{Calculates the mean over all numeric columns in the given data frame.}
\usage{
df_mean(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
The mean of the data frame.
}
\description{
Calculates the mean over all numeric columns in the given data frame.
}
\examples{
m = df_mean(test_data)
print(m)
}
