% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfStats.R
\name{df_standard_deviation}
\alias{df_standard_deviation}
\title{Calculates the standard deviation over all numeric columns in the given data frame.}
\usage{
df_standard_deviation(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
The standard deviation of the data frame.
}
\description{
Calculates the standard deviation over all numeric columns in the given data frame.
}
\examples{
standard_deviation = df_standard_deviation(test_data)
print(standard_deviation)
}
