% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfStats.R
\name{df_variance}
\alias{df_variance}
\title{Calculates the variance over all numeric columns in the given data frame.}
\usage{
df_variance(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
The variance of the data frame.
}
\description{
Calculates the variance over all numeric columns in the given data frame.
}
\examples{
variance = df_variance(test_data)
print(variance)
}
