% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfStats-package.R
\docType{package}
\name{dfStats-package}
\alias{dfStats-package}
\alias{dfStats}
\title{dfStats: Fast and elegant column and row operations for \code{data.frame} objects}
\description{
This package provides optimized functions for computing descriptive summaries
and aggregations on \code{data.frame} objects in R. Similar to the existing
\code{\link[matrixStats:matrixStats-package]{matrixStats}} package, but specifically
designed for \code{data.frame} operations with focus on both elegant and
computational efficiency.
}
\details{
dfStats

The \code{\link[base:apply]{apply}} function has several limitations when working
with \code{data.frame} objects for row and column operations:
\itemize{
\item Hard to remember the unintuitive syntax with the MARGIN parameter (1 for
rows, 2 for columns).
\item Poor computational performance from repeated function calls.
\item Memory inefficiency from creating multiple hard copies.
\item Requires manual handling of mixed column types.
}

The \code{dfStats} package provides a collection of functions for performing
efficient row-wise and column-wise computations on \code{data.frame} objects
with improved performance compared to base R operations. This package
address these issues by providing intuitive and consistently named
functions using C++ implementations for core operations to minimize hard
copies.
}
\examples{
if (!"remotes" \%in\% as.data.frame(installed.packages())$Package)
  install.packages("remotes")

#remotes::install_github("randef1ned/dfStats", upgrade = "always",
#                        build_manual = TRUE, build_vignettes = TRUE)

}
\references{
Inspired by the matrixStats package:

Henrik Bengtsson (2017). \emph{matrixStats: Functions that Apply to Rows and
Columns of Matrices (and to Vectors)}. R package version 0.52.2.
\url{https://github.com/HenrikBengtsson/matrixStats}.
}
\seealso{
Useful links:
\itemize{
\item \url{https://github.com/randef1ned/dfStats}
\item \url{https://dfStats.yinchun.su}
}
}
\author{
Yinchun Su
}
