% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfStats.R
\name{getNumbers}
\alias{getNumbers}
\title{Removes any row containing 'NA' in a numeric column and extracts all numbers.}
\usage{
getNumbers(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
A vector with the data frames numbers.
}
\description{
Removes any row containing 'NA' in a numeric column and extracts all numbers.
}
